Extract MockPOSCollectOrderPaymentAnalyticsTracker
